pyproject.toml
[project]
name = "pipecat-quickstart"
version = "0.1.0"
description = "Quickstart example for building voice AI bots with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[cartesia,daily,deepgram,fal,local-smart-turn-v3,openai,runner,silero,webrtc]",
    "pipecat-ai-cli",
    "twilio>=9.8.6",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
